Move null binary svc to storage
